WIP Do node stuff in separate container
* Seems to work fine with a lightweight image I found after a quick
search. Possibly there are even more suitable images somewhere (or we
might build a dedicated, but rather not, as those are slower in CI than
the more commonly used ones).
* Exclude node_modules and built assets from the php container. I
suppose that this decreases the load for the php-fpm container,
especially since node_modules consists of a huge amount of small files.
Though it's just an assumption and we should test this impact. If it
would be significant we can also look to do the same thing in other
directories not needed in other containers.