PLANET-6019: Empty ElasticPress host when ES is not running
Ref: https://jira.greenpeace.org/browse/PLANET-6019
Related to: https://github.com/greenpeace/planet4-docker-compose/pull/85

Setting ElasticPress host to empty stops it from trying to connect
to a server and throwing exceptions, and changes its error message
in the admin panel to an information one. It solves the last issues
of planet-6019 without touching the Search class.